+++ include/rawstor.h
@@ -8,6 +8,7 @@
 #define RAWSTOR_H
 
 #include <rawstor/list.h>
+#include <rawstor/location.h>
 #include <rawstor/object.h>
 #include <rawstor/protocol.h>
 #include <rawstor/rawio.h>